pyproject.toml
[project]
name = "rss-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "feedparser>=6.0.11",
    "python-dateutil>=2.9.0.post0",
]
